pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Elena Volkova <lenochka.volkova@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^6.2.5"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
